Move Hibernate Search Elasticsearch backend to a dedicated extension
