+++ devtools/bom-descriptor-json/pom.xml
@@ -956,6 +956,19 @@
                 </dependency>
                 <dependency>
                     <groupId>io.quarkus</groupId>
+                    <artifactId>quarkus-hibernate-search-backend-elasticsearch</artifactId>
+                    <version>${project.version}</version>
+                    <type>pom</type>
+                    <scope>test</scope>
+                    <exclusions>
+                        <exclusion>
+                            <groupId>*</groupId>
+                            <artifactId>*</artifactId>
+                        </exclusion>
+                    </exclusions>
+                </dependency>
+                <dependency>
+                    <groupId>io.quarkus</groupId>
                     <artifactId>quarkus-hibernate-search-orm-elasticsearch</artifactId>
                     <version>${project.version}</version>
                     <type>pom</type>
