+++ extensions/hibernate-search-backend-elasticsearch/deployment/src/main/java/io/quarkus/hibernate/search/backend/elasticsearch/deployment/HibernateSearchBackendElasticsearchEnabledBuildItem.java
@@ -1,0 +1,28 @@
+package io.quarkus.hibernate.search.backend.elasticsearch.deployment;
+
+import java.util.Map;
+
+import io.quarkus.builder.item.MultiBuildItem;
+import io.quarkus.hibernate.search.backend.elasticsearch.runtime.HibernateSearchBackendElasticsearchBuildTimeConfig;
+import io.quarkus.hibernate.search.backend.elasticsearch.runtime.MapperContext;
+
+public final class HibernateSearchBackendElasticsearchEnabledBuildItem extends MultiBuildItem {
+
+    private final MapperContext mapperContext;
+    private final Map<String, HibernateSearchBackendElasticsearchBuildTimeConfig> buildTimeConfig;
+
+    public HibernateSearchBackendElasticsearchEnabledBuildItem(MapperContext mapperContext,
+            Map<String, HibernateSearchBackendElasticsearchBuildTimeConfig> buildTimeConfig) {
+        this.mapperContext = mapperContext;
+        this.buildTimeConfig = buildTimeConfig;
+    }
+
+    public MapperContext getMapperContext() {
+        return mapperContext;
+    }
+
+    public Map<String, HibernateSearchBackendElasticsearchBuildTimeConfig> getBuildTimeConfig() {
+        return buildTimeConfig;
+    }
+
+}
