+++ extensions/hibernate-search-backend-elasticsearch/runtime/src/main/java/io/quarkus/hibernate/search/backend/elasticsearch/runtime/ElasticsearchVersionSubstitution.java
@@ -1,4 +1,4 @@
-package io.quarkus.hibernate.search.orm.elasticsearch.runtime;
+package io.quarkus.hibernate.search.backend.elasticsearch.runtime;
 
 import org.hibernate.search.backend.elasticsearch.ElasticsearchVersion;
