+++ extensions/hibernate-search-backend-elasticsearch/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,6 @@
+---
+artifact: ${project.groupId}:${project.artifactId}:${project.version}
+name: "Hibernate Search + Elasticsearch"
+metadata:
+  unlisted: true
+  status: "stable"
