+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/main/java/io/quarkus/hibernate/search/orm/elasticsearch/deployment/HibernateSearchElasticsearchProcessor.java
@@ -5,12 +5,9 @@ import static io.quarkus.hibernate.search.orm.elasticsearch.deployment.ClassName
 import static io.quarkus.hibernate.search.orm.elasticsearch.deployment.ClassNames.ROOT_MAPPING;
 import static io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchElasticsearchRuntimeConfig.backendPropertyKey;
 import static io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchElasticsearchRuntimeConfig.defaultBackendPropertyKeys;
-import static io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchElasticsearchRuntimeConfig.elasticsearchVersionPropertyKey;
 import static io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchElasticsearchRuntimeConfig.mapperPropertyKey;
 import static io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchElasticsearchRuntimeConfig.mapperPropertyKeys;
 
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -18,14 +15,10 @@ import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.Set;
 
 import org.hibernate.search.backend.elasticsearch.ElasticsearchVersion;
-import org.hibernate.search.backend.elasticsearch.analysis.ElasticsearchAnalysisConfigurer;
-import org.hibernate.search.backend.elasticsearch.gson.spi.GsonClasses;
-import org.hibernate.search.backend.elasticsearch.index.layout.IndexLayoutStrategy;
 import org.hibernate.search.engine.reporting.FailureHandler;
 import org.hibernate.search.mapper.orm.automaticindexing.session.AutomaticIndexingSynchronizationStrategy;
 import org.hibernate.search.mapper.pojo.work.IndexingPlanSynchronizationStrategy;
@@ -42,12 +35,10 @@ import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.annotations.BuildSteps;
 import io.quarkus.deployment.annotations.ExecutionTime;
 import io.quarkus.deployment.annotations.Record;
-import io.quarkus.deployment.builditem.ApplicationArchivesBuildItem;
 import io.quarkus.deployment.builditem.CombinedIndexBuildItem;
 import io.quarkus.deployment.builditem.DevServicesAdditionalConfigBuildItem;
 import io.quarkus.deployment.builditem.HotDeploymentWatchedFileBuildItem;
 import io.quarkus.deployment.builditem.nativeimage.NativeImageResourceBuildItem;
-import io.quarkus.deployment.builditem.nativeimage.ReflectiveClassBuildItem;
 import io.quarkus.deployment.recording.RecorderContext;
 import io.quarkus.deployment.util.JandexUtil;
 import io.quarkus.elasticsearch.restclient.common.deployment.DevservicesElasticsearchBuildItem;
@@ -58,15 +49,14 @@ import io.quarkus.hibernate.orm.deployment.integration.HibernateOrmIntegrationSt
 import io.quarkus.hibernate.orm.runtime.PersistenceUnitUtil;
 import io.quarkus.hibernate.orm.runtime.integration.HibernateOrmIntegrationRuntimeInitListener;
 import io.quarkus.hibernate.orm.runtime.integration.HibernateOrmIntegrationStaticInitListener;
-import io.quarkus.hibernate.search.orm.elasticsearch.runtime.ElasticsearchVersionSubstitution;
+import io.quarkus.hibernate.search.backend.elasticsearch.deployment.HibernateSearchBackendElasticsearchEnabledBuildItem;
+import io.quarkus.hibernate.search.backend.elasticsearch.runtime.ElasticsearchVersionSubstitution;
 import io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchElasticsearchBuildTimeConfig;
 import io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchElasticsearchBuildTimeConfigPersistenceUnit;
-import io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchElasticsearchBuildTimeConfigPersistenceUnit.ElasticsearchBackendBuildTimeConfig;
-import io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchElasticsearchBuildTimeConfigPersistenceUnit.ElasticsearchIndexBuildTimeConfig;
 import io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchElasticsearchRecorder;
 import io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchElasticsearchRuntimeConfig;
+import io.quarkus.hibernate.search.orm.elasticsearch.runtime.HibernateSearchOrmElasticsearchMapperContext;
 import io.quarkus.runtime.configuration.ConfigUtils;
-import io.quarkus.runtime.configuration.ConfigurationException;
 import io.quarkus.vertx.http.deployment.spi.RouteBuildItem;
 
 @BuildSteps(onlyIf = HibernateSearchEnabled.class)
@@ -82,7 +72,6 @@ class HibernateSearchElasticsearchProcessor {
             CombinedIndexBuildItem combinedIndexBuildItem,
             HibernateSearchElasticsearchBuildTimeConfig buildTimeConfig,
             List<PersistenceUnitDescriptorBuildItem> persistenceUnitDescriptorBuildItems,
-            BuildProducer<ReflectiveClassBuildItem> reflectiveClass,
             BuildProducer<HibernateSearchElasticsearchPersistenceUnitConfiguredBuildItem> configuredPersistenceUnits,
             BuildProducer<HibernateOrmIntegrationStaticConfiguredBuildItem> staticIntegrations,
             BuildProducer<HibernateOrmIntegrationRuntimeConfiguredBuildItem> runtimeIntegrations,
@@ -112,8 +101,6 @@ class HibernateSearchElasticsearchProcessor {
                     backendAndIndexNamesForSearchExtensions,
                     configuredPersistenceUnits, staticIntegrations, runtimeIntegrations);
         }
-
-        reflectiveClass.produce(registerReflectionForGson());
     }
 
     private static Map<String, Map<String, Set<String>>> collectPersistenceUnitAndBackendAndIndexNamesForSearchExtensions(
@@ -161,8 +148,20 @@ class HibernateSearchElasticsearchProcessor {
         }
 
         configuredPersistenceUnits
-                .produce(new HibernateSearchElasticsearchPersistenceUnitConfiguredBuildItem(persistenceUnitName, puConfig,
-                        backendNamesForIndexedEntities, backendAndIndexNamesForSearchExtensions));
+                .produce(new HibernateSearchElasticsearchPersistenceUnitConfiguredBuildItem(
+                        new HibernateSearchOrmElasticsearchMapperContext(persistenceUnitName,
+                                backendNamesForIndexedEntities, backendAndIndexNamesForSearchExtensions),
+                        puConfig));
+    }
+
+    @BuildStep
+    void enableBackend(List<HibernateSearchElasticsearchPersistenceUnitConfiguredBuildItem> enabledPUs,
+            BuildProducer<HibernateSearchBackendElasticsearchEnabledBuildItem> elasticsearchEnabled) {
+        for (HibernateSearchElasticsearchPersistenceUnitConfiguredBuildItem enabled : enabledPUs) {
+            var buildTimeConfig = enabled.getBuildTimeConfig();
+            elasticsearchEnabled.produce(new HibernateSearchBackendElasticsearchEnabledBuildItem(enabled.mapperContext,
+                    buildTimeConfig == null ? Collections.emptyMap() : buildTimeConfig.backends()));
+        }
     }
 
     @BuildStep
@@ -174,8 +173,7 @@ class HibernateSearchElasticsearchProcessor {
 
         // Some user-injectable beans are retrieved programmatically and shouldn't be removed
         unremovableBean.produce(UnremovableBeanBuildItem.beanTypes(FailureHandler.class,
-                AutomaticIndexingSynchronizationStrategy.class, IndexingPlanSynchronizationStrategy.class,
-                ElasticsearchAnalysisConfigurer.class, IndexLayoutStrategy.class));
+                AutomaticIndexingSynchronizationStrategy.class, IndexingPlanSynchronizationStrategy.class));
     }
 
     @BuildStep
@@ -212,8 +210,9 @@ class HibernateSearchElasticsearchProcessor {
                     new HibernateOrmIntegrationStaticConfiguredBuildItem(HIBERNATE_SEARCH_ELASTICSEARCH, puName)
                             .setInitListener(
                                     // we cannot pass a config group to a recorder so passing the whole config
-                                    recorder.createStaticInitListener(puName, buildTimeConfig,
-                                            configuredPersistenceUnit.getBackendAndIndexNamesForSearchExtensions(),
+                                    recorder.createStaticInitListener(
+                                            configuredPersistenceUnit.mapperContext,
+                                            buildTimeConfig,
                                             rootAnnotationMappedClassNames,
                                             integrationStaticInitListeners))
                             .setXmlMappingRequired(xmlMappingRequired));
@@ -266,119 +265,9 @@ class HibernateSearchElasticsearchProcessor {
             runtimeConfigured.produce(
                     new HibernateOrmIntegrationRuntimeConfiguredBuildItem(HIBERNATE_SEARCH_ELASTICSEARCH, puName)
                             .setInitListener(
-                                    recorder.createRuntimeInitListener(runtimeConfig, puName,
-                                            configuredPersistenceUnit.getBackendAndIndexNamesForSearchExtensions(),
-                                            integrationRuntimeInitListeners)));
+                                    recorder.createRuntimeInitListener(configuredPersistenceUnit.mapperContext,
+                                            runtimeConfig, integrationRuntimeInitListeners)));
         }
-    }
-
-    @BuildStep
-    public void processPersistenceUnitBuildTimeConfig(
-            List<HibernateSearchElasticsearchPersistenceUnitConfiguredBuildItem> configuredPersistenceUnits,
-            ApplicationArchivesBuildItem applicationArchivesBuildItem,
-            BuildProducer<NativeImageResourceBuildItem> nativeImageResources,
-            BuildProducer<HotDeploymentWatchedFileBuildItem> hotDeploymentWatchedFiles) {
-        for (HibernateSearchElasticsearchPersistenceUnitConfiguredBuildItem configuredPersistenceUnit : configuredPersistenceUnits) {
-            processPersistenceUnitBuildTimeConfig(configuredPersistenceUnit, applicationArchivesBuildItem, nativeImageResources,
-                    hotDeploymentWatchedFiles);
-        }
-    }
-
-    private void processPersistenceUnitBuildTimeConfig(
-            HibernateSearchElasticsearchPersistenceUnitConfiguredBuildItem configuredPersistenceUnit,
-            ApplicationArchivesBuildItem applicationArchivesBuildItem,
-            BuildProducer<NativeImageResourceBuildItem> nativeImageResources,
-            BuildProducer<HotDeploymentWatchedFileBuildItem> hotDeploymentWatchedFiles) {
-        String persistenceUnitName = configuredPersistenceUnit.getPersistenceUnitName();
-        HibernateSearchElasticsearchBuildTimeConfigPersistenceUnit buildTimeConfig = configuredPersistenceUnit
-                .getBuildTimeConfig();
-
-        Set<String> propertyKeysWithNoVersion = new LinkedHashSet<>();
-        Map<String, ElasticsearchBackendBuildTimeConfig> backends = buildTimeConfig != null
-                ? buildTimeConfig.backends()
-                : Collections.emptyMap();
-
-        Set<String> allBackendNames = new LinkedHashSet<>(configuredPersistenceUnit.getBackendNamesForIndexedEntities());
-        allBackendNames.addAll(backends.keySet());
-        // For all backends referenced either through @Indexed(backend = ...) or configuration...
-        for (String backendName : allBackendNames) {
-            ElasticsearchBackendBuildTimeConfig backendConfig = backends.get(backendName);
-            // ... we validate that the backend is configured and the version is present
-            if (backendConfig == null || backendConfig.version().isEmpty()) {
-                propertyKeysWithNoVersion.add(elasticsearchVersionPropertyKey(persistenceUnitName, backendName));
-            }
-            // ... we register files referenced from backends configuration
-            if (backendConfig != null) {
-                registerClasspathFileFromBackendConfig(persistenceUnitName, backendName, backendConfig,
-                        applicationArchivesBuildItem, nativeImageResources, hotDeploymentWatchedFiles);
-            }
-        }
-        if (!propertyKeysWithNoVersion.isEmpty()) {
-            throw new ConfigurationException(
-                    "The Elasticsearch version needs to be defined via properties: "
-                            + String.join(", ", propertyKeysWithNoVersion) + ".",
-                    propertyKeysWithNoVersion);
-        }
-    }
-
-    private static void registerClasspathFileFromBackendConfig(String persistenceUnitName, String backendName,
-            ElasticsearchBackendBuildTimeConfig backendConfig,
-            ApplicationArchivesBuildItem applicationArchivesBuildItem,
-            BuildProducer<NativeImageResourceBuildItem> nativeImageResources,
-            BuildProducer<HotDeploymentWatchedFileBuildItem> hotDeploymentWatchedFiles) {
-        registerClasspathFileFromIndexConfig(persistenceUnitName, backendName, null, backendConfig.indexDefaults(),
-                applicationArchivesBuildItem, nativeImageResources, hotDeploymentWatchedFiles);
-        for (Entry<String, ElasticsearchIndexBuildTimeConfig> entry : backendConfig.indexes().entrySet()) {
-            String indexName = entry.getKey();
-            ElasticsearchIndexBuildTimeConfig indexConfig = entry.getValue();
-            registerClasspathFileFromIndexConfig(persistenceUnitName, backendName, indexName, indexConfig,
-                    applicationArchivesBuildItem, nativeImageResources, hotDeploymentWatchedFiles);
-        }
-    }
-
-    private static void registerClasspathFileFromIndexConfig(String persistenceUnitName, String backendName, String indexName,
-            ElasticsearchIndexBuildTimeConfig indexConfig,
-            ApplicationArchivesBuildItem applicationArchivesBuildItem,
-            BuildProducer<NativeImageResourceBuildItem> nativeImageResources,
-            BuildProducer<HotDeploymentWatchedFileBuildItem> hotDeploymentWatchedFiles) {
-        registerClasspathFileFromConfig(persistenceUnitName, backendName, indexName, "schema-management.settings-file",
-                indexConfig.schemaManagement().settingsFile(),
-                applicationArchivesBuildItem, nativeImageResources, hotDeploymentWatchedFiles);
-        registerClasspathFileFromConfig(persistenceUnitName, backendName, indexName, "schema-management.mapping-file",
-                indexConfig.schemaManagement().mappingFile(),
-                applicationArchivesBuildItem, nativeImageResources, hotDeploymentWatchedFiles);
-    }
-
-    private static void registerClasspathFileFromConfig(String persistenceUnitName, String backendName, String indexName,
-            String propertyKeyRadical,
-            Optional<String> classpathFileOptional,
-            ApplicationArchivesBuildItem applicationArchivesBuildItem,
-            BuildProducer<NativeImageResourceBuildItem> nativeImageResources,
-            BuildProducer<HotDeploymentWatchedFileBuildItem> hotDeploymentWatchedFiles) {
-        if (!classpathFileOptional.isPresent()) {
-            return;
-        }
-        String classpathFile = classpathFileOptional.get();
-
-        Path existingPath = applicationArchivesBuildItem.getRootArchive().getChildPath(classpathFile);
-
-        if (existingPath == null || Files.isDirectory(existingPath)) {
-            //raise exception if explicit file is not present (i.e. not the default)
-            throw new ConfigurationException(
-                    "Unable to find file referenced in '"
-                            + backendPropertyKey(persistenceUnitName, backendName, indexName, propertyKeyRadical) + "="
-                            + classpathFile
-                            + "'. Remove property or add file to your path.");
-        }
-        nativeImageResources.produce(new NativeImageResourceBuildItem(classpathFile));
-        hotDeploymentWatchedFiles.produce(new HotDeploymentWatchedFileBuildItem(classpathFile));
-    }
-
-    private ReflectiveClassBuildItem registerReflectionForGson() {
-        String[] reflectiveClasses = GsonClasses.typesRequiringReflection().toArray(String[]::new);
-        return ReflectiveClassBuildItem.builder(reflectiveClasses)
-                .reason(getClass().getName())
-                .methods().fields().build();
     }
 
     @BuildStep(onlyIfNot = IsNormal.class)
