+++ extensions/hibernate-search-orm-elasticsearch/runtime/pom.xml
@@ -26,8 +26,8 @@
             <artifactId>quarkus-elasticsearch-rest-client-common</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.hibernate.search</groupId>
-            <artifactId>hibernate-search-backend-elasticsearch</artifactId>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-hibernate-search-backend-elasticsearch</artifactId>
         </dependency>
         <dependency>
             <groupId>org.hibernate.search</groupId>
