+++ extensions/hibernate-search-orm-elasticsearch/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,6 +1,6 @@
 ---
 artifact: ${project.groupId}:${project.artifactId}:${project.version}
-name: "Hibernate Search + Elasticsearch"
+name: "Hibernate Search ORM + Elasticsearch"
 metadata:
   keywords:
   - "hibernate-search-orm-elasticsearch"
@@ -12,6 +12,7 @@ metadata:
   - "hibernate-orm"
   - "hibernate-search-orm"
   - "elasticsearch"
+  - "opensearch"
   guide: "https://quarkus.io/guides/hibernate-search-orm-elasticsearch"
   categories:
   - "data"
