+++ extensions/hibernate-search-standalone-elasticsearch/deployment/src/main/java/io/quarkus/hibernate/search/standalone/elasticsearch/deployment/HibernateSearchStandaloneEnabledBuildItem.java
@@ -1,30 +1,19 @@
 package io.quarkus.hibernate.search.standalone.elasticsearch.deployment;
 
-import java.util.Map;
 import java.util.Set;
 
 import io.quarkus.builder.item.SimpleBuildItem;
+import io.quarkus.hibernate.search.standalone.elasticsearch.runtime.HibernateSearchStandaloneElasticsearchMapperContext;
 
 public final class HibernateSearchStandaloneEnabledBuildItem extends SimpleBuildItem {
 
-    private final Set<String> backendNamesForIndexedEntities;
-    private final Map<String, Set<String>> backendAndIndexNamesForSearchExtensions;
+    final HibernateSearchStandaloneElasticsearchMapperContext mapperContext;
     private final Set<String> rootAnnotationMappedClassNames;
 
-    public HibernateSearchStandaloneEnabledBuildItem(Set<String> backendNamesForIndexedEntities,
-            Map<String, Set<String>> backendAndIndexNamesForSearchExtensions,
+    public HibernateSearchStandaloneEnabledBuildItem(HibernateSearchStandaloneElasticsearchMapperContext mapperContext,
             Set<String> rootAnnotationMappedClassNames) {
-        this.backendNamesForIndexedEntities = backendNamesForIndexedEntities;
-        this.backendAndIndexNamesForSearchExtensions = backendAndIndexNamesForSearchExtensions;
+        this.mapperContext = mapperContext;
         this.rootAnnotationMappedClassNames = rootAnnotationMappedClassNames;
-    }
-
-    public Set<String> getBackendNamesForIndexedEntities() {
-        return backendNamesForIndexedEntities;
-    }
-
-    public Map<String, Set<String>> getBackendAndIndexNamesForSearchExtensions() {
-        return backendAndIndexNamesForSearchExtensions;
     }
 
     public Set<String> getRootAnnotationMappedClassNames() {
