+++ extensions/hibernate-search-standalone-elasticsearch/runtime/pom.xml
@@ -22,8 +22,8 @@
             <artifactId>quarkus-elasticsearch-rest-client-common</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.hibernate.search</groupId>
-            <artifactId>hibernate-search-backend-elasticsearch</artifactId>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-hibernate-search-backend-elasticsearch</artifactId>
         </dependency>
         <dependency>
             <groupId>org.hibernate.search</groupId>
