+++ extensions/hibernate-search-standalone-elasticsearch/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -9,6 +9,7 @@ metadata:
   - "full-text"
   - "hibernate"
   - "elasticsearch"
+  - "opensearch"
   guide: "https://quarkus.io/guides/hibernate-search-standalone-elasticsearch"
   categories:
   - "data"
